feat(43844): Refatora métodos e propriedades de Conferencia de Lancamentos
