+++ src/componentes/dres/PrestacaoDeContas/DetalhePrestacaoDeContas/GetComportamentoPorStatus.js
@@ -11,7 +11,7 @@ import {CobrancaDevolucoesPrestacaoDeContas} from "./CobrancaDevolucoesPrestacao
 import {ComentariosDeAnalise} from "./ComentariosDeAnalise";
 import {TabsArquivosDeReferencia} from "./ArquivosDeReferencia/TabsArquivosDeReferencia";
 import ArquivosDeReferenciaVisualizacaoDownload from "./ArquivosDeReferencia/ArquivosDeReferenciaVisualizacaoDownload";
-import {TabsConferenciaDeLancamentos} from "./ConferenciaDeLancamentos/TabsConferenciaDeLancamentos";
+import ConferenciaDeLancamentos from "./ConferenciaDeLancamentos";
 
 
 export const GetComportamentoPorStatus = (
@@ -62,11 +62,6 @@ export const GetComportamentoPorStatus = (
         setShowVoltarParaAnalise,
         btnSalvarDisabled,
         setBtnSalvarDisabled,
-        carregaLancamentosParaConferencia,
-        setLancamentosParaConferencia,
-        lancamentosParaConferencia,
-        contaUuid,
-        loadingLancamentosParaConferencia,
     }) => {
 
     if (prestacaoDeContas && prestacaoDeContas.status) {
@@ -226,18 +221,13 @@ export const GetComportamentoPorStatus = (
                         prestacaoDeContas={prestacaoDeContas}
                     />
 
-                    <TabsConferenciaDeLancamentos
-                        infoAta={infoAta}
-                        toggleBtnEscolheConta={toggleBtnEscolheConta}
-                        clickBtnEscolheConta={clickBtnEscolheConta}
-                        carregaLancamentosParaConferencia={carregaLancamentosParaConferencia}
-                        prestacaoDeContas={prestacaoDeContas}
-                        setLancamentosParaConferencia={setLancamentosParaConferencia}
-                        lancamentosParaConferencia={lancamentosParaConferencia}
-                        contaUuid={contaUuid}
-                        loadingLancamentosParaConferencia={loadingLancamentosParaConferencia}
+                <ConferenciaDeLancamentos
+                    infoAta={infoAta}
+                    toggleBtnEscolheConta={toggleBtnEscolheConta}
+                    clickBtnEscolheConta={clickBtnEscolheConta}
+                    prestacaoDeContas={prestacaoDeContas}
+                />
 
-                    />
                     <ComentariosDeAnalise
                         prestacaoDeContas={prestacaoDeContas}
                     />
